+++ src/frontend/apps/conversations/src/features/chat/components/SendButton.tsx
@@ -39,7 +39,9 @@ export const SendButton = ({
         <Button
           size="small"
           type="submit"
-          className="c__button--send"
+          className={
+            disabled ? 'c__button--stop c__button--send' : 'c__button--send'
+          }
           aria-label={t('Send')}
           disabled={disabled || status === 'error'}
           variant="primary"